requirements.txt
alabaster==0.7.12
asgiref==3.4.1
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
Babel==2.9.1
black==21.10b0
cached-property==1.5.2
Cerberus==1.3.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
configparser==5.1.0
coverage==6.0
crayons==0.4.0
cryptography==35.0.0
distlib==0.3.4
Django==3.2.5
django-cors-headers==3.11.0
django-extensions==3.1.3
django-extra-views==0.14.0
django-filter==21.1
django-money==2.1
django-nested-admin==3.3.3
django-smart-selects==1.5.9
django-treebeard==4.5.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
docopt==0.6.2
docutils==0.17.1
drf-writable-nested==0.6.3
entrypoints==0.3
factory-boy==3.2.1
Faker==9.8.1
flake8==4.0.1
h11==0.12.0
idna==3.3
imagesize==1.2.0
importlib-metadata==4.10.0
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.0.2
Markdown==3.3.6
MarkupSafe==2.0.1
mccabe==0.6.1
model-bakery==1.3.2
mypy-extensions==0.4.3
orderedmultidict==1.0.1
outcome==1.1.0
packaging==21.0
pathspec==0.9.0
pep517==0.12.0
Pillow==8.3.1
pip-api==0.0.26
pip-shims==0.6.0
pipreqs==0.4.11
platformdirs==2.4.0
plette==0.2.3
pluggy==1.0.0
py==1.11.0
py-moneyed==1.2
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.10.0
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.0
pytest==6.2.5
pytest-base-url==1.4.2
pytest-django==4.4.0
pytest-factoryboy==2.1.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-selenium==2.0.1
pytest-variables==1.9.0
python-dateutil==2.8.2
python-monkey-business==1.0.0
pytz==2021.1
regex==2021.11.10
requests==2.26.0
requirementslib==1.6.1
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.1.0
sortedcontainers==2.4.0
Sphinx==4.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
tenacity==6.3.1
text-unidecode==1.3
toml==0.10.2
tomli==1.2.2
tomlkit==0.8.0
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
urllib3==1.26.7
vistir==0.5.2
webdriver-manager==3.5.2
wsproto==1.0.0
yarg==0.1.9
zipp==3.7.0